pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "context-md"
version = "0.1.0"
description = "Runtime context management for AI agents - Git-native, zero dependencies"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Context.md Team", email = "context-md@example.com"}
]
keywords = ["ai", "agents", "context", "git", "automation", "devtools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Software Development :: Quality Assurance",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "keyring>=24.0.0",
    "python-docx>=1.1.0",
    "markdown-it-py>=3.0.0",
    "markdown>=3.5.0",
    "websockets>=12.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "responses>=0.24.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
export = [
    "docx2pdf>=0.1.8; platform_system=='Windows'",
    "pypandoc>=1.12",
]
agent = [
    "agent-framework>=0.1.0a0",
    "pyyaml>=6.0.0",
]
agent-anthropic = [
    "agent-framework>=0.1.0a0",
    "agent-framework-anthropic>=0.1.0a0",
    "pyyaml>=6.0.0",
]

[project.scripts]
context-md = "context_md.cli:main"

[project.urls]
Homepage = "https://github.com/jnPiyush/ContextMD"
Documentation = "https://github.com/jnPiyush/ContextMD/blob/main/docs/architecture/ARCHITECTURE-CONTEXT-MD.md"
Repository = "https://github.com/jnPiyush/ContextMD"
Issues = "https://github.com/jnPiyush/ContextMD/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["context_md*"]

[tool.setuptools.package-data]
context_md = ["static/*"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.isort]
known-first-party = ["context_md"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["context_md"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
